Cast from float to int as needed

+++ src/main/java/slimeknights/tconstruct/world/client/SlimeColorizer.java
@@ -22,7 +22,7 @@ public class SlimeColorizer {
   });
 
   public static final float LOOP = 256; // after how many blocks the pattern loops
-  public static final BlockPos LOOP_OFFSET = new BlockPos(SlimeColorizer.LOOP / 2, 0, SlimeColorizer.LOOP / 2);
+  public static final BlockPos LOOP_OFFSET = new BlockPos((int) (SlimeColorizer.LOOP / 2), 0, (int) (SlimeColorizer.LOOP / 2));
 
   /**
    * Updates the colors for the given type
